clean up SB
