+++ .storybook/main.js
@@ -8,13 +8,5 @@ module.exports = {
     "@storybook/addon-essentials",
     "@storybook/addon-a11y",
     "storybook-dark-mode",
-    // {
-    //   name: "@storybook/addon-postcss",
-    //   options: {
-    //     postcssLoaderOptions: {
-    //       implementation: require("postcss"),
-    //     },
-    //   },
-    // },
   ]
 }
